+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.1.0 - First Release
-* Every feature added
-* Every bug fixed
+* color code block: solid color
+* color code block: gradient color